Update IKOS formula to install version 3.2 (#9).
This commit updates the formula and checksum to point to release 3.2 of IKOS.
The installation process also changes, and now requires using pip after cmake
to install IKOS as a python package.

The new release of IKOS requires python 3, so we adjust the dependency
accordingly. IKOS has a NOSA license, not MIT, so we adjust the license as
well.